Add Help menu and update main form title
A new "Help" menu has been added to the menu strip in the MainForm.cs file.
Within the "Help" menu, an "About" menu item has been added. When selected, it displays a message box with information about the Acorn BASIC Deobfuscator tool, including the author, contributors, and a link to the GitHub repository.
The title of the main form has been changed from "Acorn BASIC Detokenizer" to "Acorn BASIC Deobfuscator".